Bugfix: dealt with the case rows = 0, add suffixes to output file, add percentage of progress

+++ cpp/src/include/tweets_spatial_analysis.h
@@ -9,6 +9,6 @@ using namespace arma;
 using namespace std;
 
 imat APD(imat A);
-void build_overlap_matrix(string input_file, string output_file, int rows = 0);
+void build_overlap_matrix(string input_file, string output_file, long rows = 0);
 
 #endif
